update by BE 9

+++ BE/src/main/java/org/example/be/service/PostService.java
@@ -48,9 +48,14 @@ public class PostService {
     }
 
     // Delete
-    public void deletePost(Integer id) {
-        postRepository.deleteById(id);
+    public boolean deletePost(Integer id) {
+        if (postRepository.existsById(id)) {
+            postRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
+
 
     // --- For You ---
     public List<Post> getPostsForYou(Integer memberId) {
@@ -68,6 +73,15 @@ public class PostService {
 
     // --- Latest Posts ---
     public List<Post> getLatestPosts(int limit) {
-        return postRepository.findLatestPosts(PageRequest.of(0, limit));
+        return postRepository.findLatestPosts(PageRequest.of(0, 8));
+    }
+    // Lấy 8 post vehicle mới nhất
+    public List<Post> getLatestVehiclePosts(int limit) {
+        return postRepository.findLatestVehiclePosts(PageRequest.of(0, 8));
+    }
+
+    // Lấy 8 post battery mới nhất
+    public List<Post> getLatestBatteryPosts(int limit) {
+        return postRepository.findLatestBatteryPosts(PageRequest.of(0, 8));
     }
 }
